feat: Implement draft management and save functionality across course content components
