+++ src/app/(main)/course-learning/[courseId]/learning/layout.tsx
@@ -7,7 +7,8 @@ import Footer from '@/components/footer';
 import { Toaster } from 'sonner';
 import { Geist, Geist_Mono } from 'next/font/google';
 import '../../../../globals.css';
-import { DraftProvider, useDraft } from '@/app/context/draftContext';
+import { useDraft } from '@/app/context/draftContext';
+import { DraftProvider } from '@/app/context/draftContext'; // ✅ import this!
 import DraftDialog from "@/components/common/DraftDialog";
 
 const geistSans = Geist({ variable: '--font-geist-sans', subsets: ['latin'] });
@@ -20,7 +21,6 @@ function LayoutContent({ children }: { children: React.ReactNode }) {
   const profilePage = pathname === '/profile';
   const paymentPage = pathname.startsWith('/payment');
 
-  // --- Draft navigation logic ---
   const { dirtyAssignments, saveAllDrafts, clearDrafts } = useDraft?.() || {};
   const [pendingNav, setPendingNav] = useState<string | null>(null);
   const [showDraftModal, setShowDraftModal] = useState(false);
@@ -56,11 +56,9 @@ function LayoutContent({ children }: { children: React.ReactNode }) {
       if (clearDrafts) clearDrafts();
     }
   };
-  // --- End draft navigation logic ---
 
   return (
     <div className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col min-h-screen overflow-x-hidden`}>
-      {/* Draft confirmation modal */}
       <DraftDialog
         open={showDraftModal}
         onOpenChange={setShowDraftModal}
@@ -75,6 +73,7 @@ function LayoutContent({ children }: { children: React.ReactNode }) {
   );
 }
 
+// ✅ Wrap in DraftProvider here
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <DraftProvider>
